+++ main_api/src/main/java/com/cinema/main_api/controller/CartController.java
@@ -19,7 +19,7 @@ public class CartController {
 
     @GetMapping("")
     @ApiOperation("根据用户id查询购物车")
-    public List<CartVo> list(@RequestParam(name = "uid") Integer uid){
+    public List<CartVo> list(@RequestParam(name = "uid") Long uid){
         return cartService.findAllByUserId(uid);
     }
 }
