+++ read_api/src/main/java/com/cinema/read_api/mapper/ReadFilmMapper.java
@@ -9,7 +9,7 @@ import java.util.Map;
 
 @Mapper
 public interface ReadFilmMapper {
-    Film selectById(Integer id);
+    Film selectById(Long id);
