+++ read_api/src/main/java/com/cinema/read_api/service/ReadFilmService.java
@@ -18,6 +18,6 @@ public interface ReadFilmService {
 
 
 
-    Film findById(Integer id);
+    Film findById(Long id);
 
 }
