+++ read_api/src/main/java/com/cinema/read_api/service/ReadOrderService.java
@@ -11,5 +11,5 @@ public interface ReadOrderService {
 
 
 
-    List<OrderVo> findOrderVoList(Integer Uid);
+    List<OrderVo> findOrderVoList(Long Uid);
 }
