+++ read_api/src/main/java/com/cinema/read_api/service/impl/ReadFilmServiceImpl.java
@@ -24,7 +24,7 @@ public class ReadFilmServiceImpl implements ReadFilmService {
     }
 
     @Override
-    public Film findById(Integer id) {
+    public Film findById(Long id) {
         return readFilmMapper.selectById(id);
     }
 }
